Avoid a diff in `ifcfg-*` files
Those files are changed by Wicked, e.g. when calling `wicked reload` on the
interface. Wicked changes the mode to something else and adds the `ZONE`.

We can simply leave out the mode to keep the mode Wicked uses.

While it is not necessary to specify the zone here explicitly, we can just
add it to avoid a diff (especially because the zone will always be
trusted).