Make default_interface grain independent of SUSE specifics
For salt-states-openqa!1316 we need a solid and reliable way to find the
default interface of a machine. It also fixes an existing usage in
openqa/worker.sls where we use it to set the net.ipv6.conf.*.accept_ra
sysctl. It is based on a solution from https://stackoverflow.com/a/6556951