Try to avoid useless salt timeouts with insane 1h value
Often some hosts timeout trying to apply the salt states but it seems
that jobs finish on the salt minions in the background. So instead of
causing random timeouts on various hosts this commit tries to apply a
much higher timeout of 1h vs. the former value of 3m. Keep in mind that
the default salt timeout is actually 5s(!).

Related progress issue: https://progress.opensuse.org/issues/175407