Avoid broken state yaml if dns in template fails
If the requested domain has no entry in DNS we have an empty ip_list
variable which will result in a broken yaml-configuration for the state
making use of that variable. The domain itself is implicitly checked by
this as well (if it is empty, dns will be empty too, resulting in beeing
skipped to render completely).