Ensure correct permissions of `openqa-asset-config.inc`
Salt created this file so only the owner/root can access it which is
wrong as NGINX cannot access the file then.

According to the documentation "The default mode for new files and directories
corresponds umask of salt process" so it makes probably sense to specify
the mode explicitly.

I also added the mode to existing files so they are created with the
right permissions if we would use Salt to setup a host from scratch.

Note that we set the mode like this already in many other places.

Related ticket: https://progress.opensuse.org/issues/175902