Switch back to sufficiently up-to-date velociraptor from official repos
In the meantime package versions in openSUSE Tumbleweed and Leap seem to
be in a better state, e.g. after
https://build.opensuse.org/request/show/1238560 having 0.7.0.4 included.
This commit removes the addition of the additional internal repository
and just uses the proper official repositories again.

Actual removal of repositories from the production hosts needs to be
done manually.

Related progress issue: https://progress.opensuse.org/issues/159060