Fix timeouts on combination of simple as well as complex commands
In case of heavy salt commands like applying the high state we had
problems with too low timeouts which we fixed by setting an insane 1h
timeout in the salt master config. However that means that if a node is
not responsive even on a simple "test.ping" we would only get a response
after 1h. This commit tries to solve that by setting back the global
timeout to the value we had before which is 90s and only setting an
explicit 1h timeout on the heavy "salt.highstate" command. Following the
same principle if we encounter more specific cases we can apply a
timeout with the "-t" command line parameter.

Related progress issue: https://progress.opensuse.org/issues/177366