Fix systemd services alert after 4429f893f
Commit 4429f893f changed the reducer in the alert condition from `max()` to
`last()`. This makes also sense (see the message of this commit) and
reverting 4429f893f would lead to too many alerts again. However, with
4429f893f we have now to few alerts.

When coming up with 4429f893f I overlooked that one data point in our query
for the number of failing systemd services does *not* represent the total
number of failing systemd services at this point in time. The query uses
the `sum()` aggregation so one might think that is the case. However, the
different hosts do not report the number of failing systemd services
in-sync with each other. So one data point in our query only represents the
number of failing systemd services on the hosts that reported their figure
around that time.

We already group data points by a certain time interval so multiple data
points are combined. However, this interval is so far computed by Grafana
using the `$__interval` variable. By increasing this interval to 70 seconds
this change groups more data points together. One data point will still
only represent the number of failing systemd services on hosts that
reported their figure around that time. However, now "around that time" is
a rather big windows of over a minute - and each host seems to report one
figure per minute. So now one data point will effectively represent the
actual total of failed systemd services.

I chose the interval of 70 seconds because it is slightly more than the
interval in which hosts are supposed to report a new number of systemd
services. I also playd around with the interval in the explore view on
Grafana. With `$_interval` there are many "dots" with zero value present
and a few non-zero spikes. With 70 seconds there are no more zero dots.
Instead, there is just a line showing the value of the previous spikes.

I suppose this will fix the alert because now `last() > 0` really means
"total sum of systemd services > 0".

Related ticket: https://progress.opensuse.org/issues/178015