Separate away templated alerts from dashboard definitions
Several grafana versions ago, all alerts were closely coupled to a
panel/dashboard. When we migrated to unified alerting with provisioned
alerts, we kept them close to the dashboard definition because it made
sense. By now, dashboards and alerts are only loosely coupled and we
have more (newer and unrelated) alerts now so it makes sense to group
them together in their own folder.