Prevent potentially corrupting ini-file updates with "file.managed"
As we changed to use a dedicate drop-in configuration file for the
openQA server we can now also use the much more efficient "file.managed"
state instead of "ini.options_present" to write the complete file in one
go. I tested this in our CI test container environment with running

```
notifywait --monitor --recursive /etc/openqa/ &
```

to monitor file operations and then call

```
salt-call --state-output=changes --local state.apply openqa.server 2>&1 | grep -v '\(DeprecationWarning\|Result: Clean\)'
```

This showed that the "ini.options_present"-approach triggers around 50
(!) open-events on the .ini file for a noop which reduced to a single
open-event for a noop using "file.managed".

The additional change in behaviour is that manual changes to the ini
file will be overwritten on the next application of a salt high state
however that is irrelevant as we can use another drop-in file for any
such changes.

Related progress issues:
* https://progress.opensuse.org/issues/177276
* https://progress.opensuse.org/issues/176250