Make kdump-notify wait for postfix on bootup
On 2025-05-30 we observed a reboot of a machine with the
kdump-notify.service failing because the used mail server responded with
"451 4.7.1 Please try again later". According to systemd, the
kdump-notify.service was started slightly before postfix and most likely
postfix was still starting up and not ready to receive mails yet. This
makes the service wait for postfix to hopefully signal readiness over
systemd before attempting to enqueue a mail. A manual test on arm-1
seemed to work and with this change applied the mail got sent after
another reboot without any error message or failing services.