Workaround missing ipv4 setting in telegraf inputs.ping plugin
This is a workaround for a missing feature switch in our used telegraf
versions. It uses jinja2 and salts dnsutil-function to resolve the IPv4
and IPv6 for a specified host and produces the following telegraf
metrics:

```
> ping,host=workstation,protocol_version=6,url=2a07:de40:b203:12:0:ff:fe4f:7c2b average_response_ms=8.772,maximum_response_ms=8.772,minimum_response_ms=8.772,packets_received=1i,packets_transmitted=1i,percent_packet_loss=0,result_code=0i,standard_deviation_ms=0,ttl=56i 1750713503000000000
> ping,host=workstation,protocol_version=4,url=10.145.10.207 average_response_ms=8.799,maximum_response_ms=8.799,minimum_response_ms=8.799,packets_received=1i,packets_transmitted=1i,percent_packet_loss=0,result_code=0i,standard_deviation_ms=0,ttl=57i 1750713503000000000
```

To match these metrics with our usual conventions, it maps back these ip
addresses into a human readable string all based on the previous DNS
lookups resulting in clean `url=openqa.suse.de`-tags.