Store groupless jobs shorter again
After the introduction of being able to keep jobs in the database after
results have been cleaned up we store groupless jobs now significantly
longer than before.

Before, these jobs were cleaned up after 15 days or 50 days if
important. Now the database entry is kept for 730 days or indefinitely
if the job is important. That's maybe a bit more than what we need and
want for jobs that are not in any group. So this change reduces this to
100 and 300 days. That's still more than before so we'll still benefit
from keeping jobs longer in the database but also more reasonable and
we don't have to be surprised finding very old jobs like in
https://progress.opensuse.org/issues/189174 again.